# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users": {
      "post": {
        "operationId": "UserAdmin_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelopeDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 201
                        },
                        "data": {
                          "$ref": "#/components/schemas/AuthUserDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request data is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated actor does not have access.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "409": {
            "description": "The request conflicts with the current resource state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "tags": [
          "UserAdmin"
        ]
      },
      "get": {
        "operationId": "UserAdmin_listUsers",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "example": "staff@example.com",
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "schema": {
              "example": "STAFF",
              "type": "string",
              "enum": [
                "STAFF",
                "ADMIN"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "example": "ACTIVE",
              "type": "string",
              "enum": [
                "ACTIVE",
                "LOCKED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelopeDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/AuthUserDto"
                          }
                        },
                        "meta": {
                          "$ref": "#/components/schemas/PaginationMetaDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated actor does not have access.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "tags": [
          "UserAdmin"
        ]
      }
    },
    "/api/v1/users/{id}": {
      "patch": {
        "operationId": "UserAdmin_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelopeDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "data": {
                          "$ref": "#/components/schemas/AuthUserDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request data is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated actor does not have access.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "409": {
            "description": "The request conflicts with the current resource state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "tags": [
          "UserAdmin"
        ]
      }
    },
    "/api/v1/users/{id}/status": {
      "patch": {
        "operationId": "UserAdmin_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccountStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelopeDto"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "data": {
                          "$ref": "#/components/schemas/AuthUserDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request data is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated actor does not have access.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "409": {
            "description": "The request conflicts with the current resource state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "tags": [
          "UserAdmin"
        ]
      }
    }
  },
  "info": {
    "title": "Homestay Booking Management System API",
    "description": "HTTP contract for the HBMS customer and management applications.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "SuccessEnvelopeDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "Thanh cong."
          },
          "data": {
            "type": "object",
            "nullable": true,
            "description": "Response payload. The concrete schema is defined per route."
          },
          "path": {
            "type": "string",
            "example": "/api/v1/resource"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-07-27T04:00:00.000Z",
            "format": "date-time"
          },
          "requestId": {
            "type": "string",
            "example": "f75cf4ac-6958-4d59-8a4f-183abfeaed15"
          }
        },
        "required": [
          "success",
          "statusCode",
          "message",
          "data",
          "path",
          "timestamp",
          "requestId"
        ]
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "COMMON_VALIDATION_FAILED",
          "COMMON_UNAUTHORIZED",
          "COMMON_FORBIDDEN",
          "COMMON_NOT_FOUND",
          "COMMON_CONFLICT",
          "COMMON_RATE_LIMITED",
          "COMMON_PAYLOAD_TOO_LARGE",
          "COMMON_UNSUPPORTED_MEDIA_TYPE",
          "COMMON_SERVICE_UNAVAILABLE",
          "COMMON_INTERNAL_ERROR",
          "CUSTOMER_INITIAL_PASSWORD_ALREADY_CONFIGURED",
          "CUSTOMER_EMAIL_IN_USE",
          "CUSTOMER_PHONE_IN_USE",
          "CUSTOMER_CURRENT_PASSWORD_INVALID",
          "CUSTOMER_PASSWORD_REUSE_NOT_ALLOWED",
          "AMENITY_NAME_ALREADY_EXISTS",
          "AMENITY_IN_USE",
          "BOOKING_GUEST_CAPACITY_EXCEEDED",
          "BOOKING_CUSTOMER_CONTACT_REQUIRED",
          "BOOKING_ACTIVE_UNPAID_LIMIT_REACHED",
          "BOOKING_HELD_NIGHTS_LIMIT_REACHED",
          "BOOKING_ROOM_NOT_BOOKABLE",
          "BOOKING_CHECKIN_IN_PAST",
          "BOOKING_CHECKIN_TOO_FAR",
          "BOOKING_DATE_RANGE_INVALID",
          "BOOKING_STAY_TOO_LONG",
          "BOOKING_TOTAL_LIMIT_EXCEEDED",
          "BOOKING_ROOM_UNAVAILABLE",
          "BOOKING_CREATE_CONFLICT",
          "BOOKING_CANCELLATION_REASON_REQUIRED",
          "BOOKING_REFUND_PENDING",
          "BOOKING_TRANSITION_NOT_ALLOWED",
          "BOOKING_CONFIRMATION_REQUIRES_PAYMENT",
          "BOOKING_CHECKIN_REQUIRES_PAYMENT",
          "BOOKING_CHECKIN_OUTSIDE_STAY_WINDOW",
          "BOOKING_ROOM_NOT_FOUND",
          "BOOKING_ROOM_NOT_READY",
          "BOOKING_CANCELLATION_ALREADY_PAID",
          "BOOKING_CANCELLATION_NOT_ALLOWED",
          "PAYMENT_REFUND_REJECTED",
          "PAYMENT_REFUND_NOT_ALLOWED",
          "PAYMENT_REFUND_OUTCOME_UNKNOWN",
          "PAYMENT_IDEMPOTENCY_KEY_CONFLICT"
        ]
      },
      "ErrorEnvelopeDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "errorCode": {
            "example": "COMMON_VALIDATION_FAILED",
            "allOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              }
            ]
          },
          "message": {
            "example": "Du lieu khong hop le.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          },
          "fieldErrors": {
            "type": "object",
            "description": "Validation errors keyed by request field.",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "object",
                "required": [
                  "errorCode"
                ],
                "properties": {
                  "errorCode": {
                    "type": "string",
                    "enum": [
                      "COMMON_VALIDATION_FAILED",
                      "COMMON_UNAUTHORIZED",
                      "COMMON_FORBIDDEN",
                      "COMMON_NOT_FOUND",
                      "COMMON_CONFLICT",
                      "COMMON_RATE_LIMITED",
                      "COMMON_PAYLOAD_TOO_LARGE",
                      "COMMON_UNSUPPORTED_MEDIA_TYPE",
                      "COMMON_SERVICE_UNAVAILABLE",
                      "COMMON_INTERNAL_ERROR",
                      "CUSTOMER_INITIAL_PASSWORD_ALREADY_CONFIGURED",
                      "CUSTOMER_EMAIL_IN_USE",
                      "CUSTOMER_PHONE_IN_USE",
                      "CUSTOMER_CURRENT_PASSWORD_INVALID",
                      "CUSTOMER_PASSWORD_REUSE_NOT_ALLOWED",
                      "AMENITY_NAME_ALREADY_EXISTS",
                      "AMENITY_IN_USE",
                      "BOOKING_GUEST_CAPACITY_EXCEEDED",
                      "BOOKING_CUSTOMER_CONTACT_REQUIRED",
                      "BOOKING_ACTIVE_UNPAID_LIMIT_REACHED",
                      "BOOKING_HELD_NIGHTS_LIMIT_REACHED",
                      "BOOKING_ROOM_NOT_BOOKABLE",
                      "BOOKING_CHECKIN_IN_PAST",
                      "BOOKING_CHECKIN_TOO_FAR",
                      "BOOKING_DATE_RANGE_INVALID",
                      "BOOKING_STAY_TOO_LONG",
                      "BOOKING_TOTAL_LIMIT_EXCEEDED",
                      "BOOKING_ROOM_UNAVAILABLE",
                      "BOOKING_CREATE_CONFLICT",
                      "BOOKING_CANCELLATION_REASON_REQUIRED",
                      "BOOKING_REFUND_PENDING",
                      "BOOKING_TRANSITION_NOT_ALLOWED",
                      "BOOKING_CONFIRMATION_REQUIRES_PAYMENT",
                      "BOOKING_CHECKIN_REQUIRES_PAYMENT",
                      "BOOKING_CHECKIN_OUTSIDE_STAY_WINDOW",
                      "BOOKING_ROOM_NOT_FOUND",
                      "BOOKING_ROOM_NOT_READY",
                      "BOOKING_CANCELLATION_ALREADY_PAID",
                      "BOOKING_CANCELLATION_NOT_ALLOWED",
                      "PAYMENT_REFUND_REJECTED",
                      "PAYMENT_REFUND_NOT_ALLOWED",
                      "PAYMENT_REFUND_OUTCOME_UNKNOWN",
                      "PAYMENT_IDEMPOTENCY_KEY_CONFLICT"
                    ]
                  },
                  "message": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "details": {
            "type": "object",
            "description": "Safe structured context that can help a client recover.",
            "additionalProperties": false,
            "properties": {
              "retryable": {
                "type": "boolean",
                "example": false
              },
              "limit": {
                "type": "integer",
                "example": 3,
                "minimum": 0
              },
              "maxAdvanceDays": {
                "type": "integer",
                "example": 365,
                "minimum": 0
              },
              "maxStayNights": {
                "type": "integer",
                "example": 30,
                "minimum": 0
              }
            }
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "path": {
            "type": "string",
            "example": "/api/v1/resource"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-07-27T04:00:00.000Z",
            "format": "date-time"
          },
          "requestId": {
            "type": "string",
            "example": "f75cf4ac-6958-4d59-8a4f-183abfeaed15"
          }
        },
        "required": [
          "success",
          "statusCode",
          "errorCode",
          "message",
          "error",
          "path",
          "timestamp",
          "requestId"
        ]
      },
      "PaginationDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1,
            "minimum": 1
          },
          "limit": {
            "type": "number",
            "example": 20,
            "minimum": 1
          },
          "total": {
            "type": "number",
            "example": 150,
            "minimum": 0
          },
          "totalPages": {
            "type": "number",
            "example": 8,
            "minimum": 0
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "totalPages"
        ]
      },
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "pagination": {
            "$ref": "#/components/schemas/PaginationDto"
          }
        },
        "required": [
          "pagination"
        ]
      },
...
      "AuthUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "format": "uuid"
          },
          "fullName": {
            "type": "string",
            "example": "Tran Thi B"
          },
          "email": {
            "type": "string",
            "example": "admin@example.com"
          },
          "phone": {
            "type": "string",
            "example": "+84987654321",
            "nullable": true
          },
          "role": {
            "type": "string",
            "example": "ADMIN",
            "enum": [
              "STAFF",
              "ADMIN"
            ]
          },
          "status": {
            "type": "string",
            "example": "ACTIVE",
            "enum": [
              "ACTIVE",
              "LOCKED"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-07-27T04:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-07-27T04:00:00.000Z"
          }
        },
        "required": [
          "id",
          "fullName",
          "email",
          "phone",
          "role",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "UpdateAccountStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "LOCKED"
            ],
            "example": "LOCKED"
          }
        },
        "required": [
          "status"
        ]
      },
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "example": "Nguyen Van Staff",
            "maxLength": 120
          },
          "email": {
            "type": "string",
            "example": "staff@example.com",
            "maxLength": 160
          },
          "phone": {
            "type": "string",
            "example": "0901234567",
            "nullable": true
          },
          "password": {
            "type": "string",
            "example": "StrongPassword123!",
            "maxLength": 72,
            "minLength": 8,
            "writeOnly": true
          },
          "role": {
            "type": "string",
            "description": "Only STAFF accounts can be issued through this endpoint.",
            "enum": [
              "STAFF"
            ],
            "example": "STAFF"
          }
        },
        "required": [
          "fullName",
          "email",
          "password"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "example": "Nguyen Van Staff",
            "maxLength": 120
          },
          "email": {
            "type": "string",
            "example": "staff@example.com",
            "maxLength": 160
          },
          "phone": {
            "type": "string",
            "example": "0901234567",
            "nullable": true
          },
          "password": {
            "type": "string",
            "example": "UpdatedPassword456!",
            "maxLength": 72,
            "minLength": 8,
            "writeOnly": true
          },
          "role": {
            "type": "string",
            "enum": [
              "STAFF"
            ],
            "example": "STAFF"
          }
        }
      }
    }
  }
}